+++ generator/lib/src/realm_model_info.dart
@@ -104,6 +104,7 @@ class RealmModelInfo {
               if (f.name != f.realmName) 'mapTo': f.realmName,
               if (f.optional) 'optional': f.optional,
               if (f.isPrimaryKey) 'primaryKey': f.isPrimaryKey,
+              if (f.indexed) 'indexed': f.indexed,
               if (f.realmType == RealmPropertyType.object) 'linkTarget': f.basicRealmTypeName,
               if (f.realmType == RealmPropertyType.linkingObjects) ...{
                 'linkOriginProperty': f.linkOriginProperty!,